+++ skrub/datasets/__init__.py
@@ -5,11 +5,13 @@ Datasets: fetching real-world data and generating synthetic data.
 Skrub bundles ready-to-use datasets (fetched from the web on first use)
 and synthetic generators for examples and experimentation.
 
-Downloaded datasets are cached in the SKB_DATA_DIRECTORY (default: ``~/.skrub_data``).
+Downloaded datasets are cached in the SKB_DATA_DIRECTORY environment variable
+(default: ``~/.skrub_data``).
 
-**Fetching** functions download real-world datasets for regression or
-classification.
-**Generating** helpers create small synthetic dataframes for quick demos.
+**Fetching** functions are defined in _fetching.py and start with `fetch_`.
+They download real-world datasets for regression or classification.
+**Generating** helpers are defined in _generating.py and start with `make_` or `toy_`.
+They create small synthetic dataframes for quick demos.
 """
 
 from ._fetching import (
